\name{Si}
\alias{Si}
\title{Moment Estimators associated with subject-specific latent variable}
\usage{
  Si(Beta, Zi, X)
}
\value{
  numeric vector
}
\description{
  \eqn{S_i = Z_i e^{X_i \beta}} The \code{Zi} and \code{X}
  are subject-specific latent varialbe and time-independent
  covariate associated with ordered \code{y}
}

