\name{SSi}
\alias{SSi}
\title{Moment Estimators associated with subject-specific latent variable}
\usage{
  SSi(Beta, Zi, X)
}
\value{
  numeric vector
}
\description{
  \eqn{SS_i = \sum_{j=1}^{i} S_j}. The \code{Zi} and
  \code{X} are subject-specific latent varialbe and
  time-independent covariate associated with ordered
  \code{y}
}
\seealso{
  \code{\link{Si}}
}

